PuzzleVision Inital Commit with Gitignore.
Removed all of the project files as they were cluttering up the
Repo.

So far there is the ability to connect to a COM port, but it is not
pulling the properties off of the GUI just yet; that will be the next
part.

There is a primative GUI at the moment that is able to handle faults
within the system, and at the same time continue running and logging
events that are received.

PuzzleVision